clean project structure and add spring boot

+++ src/pages/rockpage/rockSwingStock/swingStockCard.jsx
@@ -1,56 +1,42 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
-import { deleteStock } from "../../../firebase/services/rockService";
-import { useNavigate } from "react-router-dom";
+import { getAuth } from "firebase/auth";
+import { userProfileData } from "../../../firebase/services/userServiceFb";
+import { onAuthStateChanged } from "firebase/auth";
+import { deleteSwingStock } from "../../service/rockService";
 
 function SwingStockCard({data}){
-const navigate = useNavigate();
 
-const deleteItem = async ()=>{
- const res = await deleteStock(data.id);
- if(res == true){
-  console.log("delete success")
- navigate("/result" , {state:{code:200 , value:true , next:"/rock"}})
- }else{
- navigate("/result" , {state:{code:401 , value:false , next:"/rock"}})
- }
+const [user,setUser] = useState("");
+
+// delete stock 
+const deleteStock = ()=>{
+  const res = deleteSwingStock(data.id)
 }
 
+// chek user roles red
+useEffect(()=>{
+const auth = getAuth();
+    const unsubscribe = onAuthStateChanged(auth,async (user) => {
+      if (user) {
+       const userdata = await userProfileData(user)
+       setUser(userdata.role);
+      } 
+    });
+    return ()=> unsubscribe();
+},[])
+
+
   return (
-    <div className={`rockcard h-50 min-w-60 max-w-60 p-3  text-black rounded-2xl shadow-md hover:shadow-2xl transition-all duration-300 flex flex-col justify-between border border-white/20 ${parseInt(data.rockAnalysis.currentPrice) > parseInt(data.rockAnalysis.minTarget) ? "bg-red-200" : "bg-green-200"}`}>
-      {/* Header */}
-      <div className="flex mb-2 justify-between items-center">
-        <h1 className="font-bold text-lg "></h1>
-        <div className=" w-full  backdrop-blur-sm h-8  rounded-full flex items-center justify-between flex-row-reverse">
-          <i className="fa-solid fa-chart-line  text-sm"></i>
-          <i className="fa-solid fa-trash cursor-pointer" onClick={deleteItem}></i>
-        </div>
-      </div>
-
-      {/* Price Section */}
-      <div className="flex flex-col items-center gap-2">
-        <div className="bg-white shadow-lg rounded-2xl px-6 py-2">
-          <p className=" font-bold text-gray-800">
-            {data.stockName}
-          </p>
-        </div>
-      </div>
-
-      {/* Footer */}
-      <div className="space-y-2 flex justify-center items-center flex-col w-full mt-4">
-        <p className="text-[10px]  text-center leading-4">
-          For educational purposes only. Please make your own investment
-          decisions.
-        </p>
-        <Link state={data}  to={"/swingStockDetails"}><p className="cursor-pointer"><i class="fa-solid fa-arrow-right"></i></p></Link>
-
-        <div className="border-t border-white/20 mb-2 flex justify-center">
-          <p className="text-[10px]">
-           {data.date}
-          </p>
-        </div>
-      </div>
+   <div className="swing-stock relative shadow-md duration-300 hover:shadow-xl  flex flex-col items-center dark:text-white h-30 min-w-60 max-w-60 rounded-2xl p-2 bg-white dark:bg-zinc-800">
+    {/* delete-button */}
+    {/* <div className={`${user == "admin" ? "" : "hidden"}`}><i className="fa-solid absolute top-2 left-2 fa-trash cursor-pointer" onClick={deleteItem}></i></div> */}
+    <div className="name h-10 flex justify-center items-center font-extrabold sm:font-bold">
+      <h1>{data.stockName}</h1>
     </div>
+    <p className="text-[10px] mb-3 text-gray-400 text-center">Educational purposes only. Please trade at your own risk. first learn no earn</p>
+     <Link state={data} to={"/swingStockDetails"}><p className="dark:text-white bg-gray-200 dark:bg-zinc-900  rounded-[10px] p-1 px-4 cursor-pointer text-[14px]">explore </p></Link>
+   </div>
   );
 }
